requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
firebase-admin==6.1.0
python-multipart==0.0.6
smtplib==0.0.1
pydantic-settings>=2.0.0 
email-validator>=1.3.0
sendgrid==6.10.0